Fix Linux/macOS CI: guard Windows-only code, add missing std headers
The Linux and macOS GitHub Actions jobs had been red for many commits as
Windows-only code accumulated. Make the example build cleanly for
GCC/AppleClang under -std=c++17 -Werror -Wall -Wpedantic -Wextra:

- TextEditor.cpp: include <cstring> for std::strlen/std::memcmp (MSVC pulled
  it in transitively, libstdc++/libc++ do not); fix an int/size_t
  sign-compare in rebuildFoldRanges.
- TextDiff.cpp: drop the unconditional <Windows.h> include (it was unused)
  and the now-orphaned NOMINMAX define.
- example/main.cpp: include <algorithm>/<cctype> for std::transform and
  std::tolower; guard the MSVC-only fopen_s with a std::fopen fallback.
- example/editor.cpp: add <cstdlib>/<cstring>/<cstdint>; define
  _DEFAULT_SOURCE on non-Windows so glibc exposes popen/pclose under strict
  c++17, and map _popen/_pclose to popen/pclose for the unguarded call sites;
  give get_module_path POSIX (readlink /proc/self/exe) and macOS
  (_NSGetExecutablePath) branches; guard the Windows-SDK registry probe;
  pair stb_image's MSVC "#pragma warning" with a GCC/Clang diagnostic
  equivalent; mark an error_code [[maybe_unused]] on the non-Windows path;
  fix two // comments ending in a backslash (-Wcomment).
- example/editor.h: same -Wcomment fix.

Each translation unit was verified to compile clean with g++ 14
-std=c++17 -Werror -Wall -Wpedantic -Wextra (main.cpp excepted, as it needs
the CMake-generated SDL config; its changes are header/guard-only).

+++ TextEditor.cpp
@@ -30,7 +30,7 @@
 #include <iterator>
 #include <memory>
 #include <string>
-#include <string.h>
+#include <cstring>
 #include <string_view>
 #include <unordered_map>
 #include <unordered_set>
@@ -3823,7 +3823,7 @@ void TextEditor::Folder::rebuildFoldRanges(Document& document)
 		const std::string text = document.getLineText(line);
 
 		// Braces { }
-		for (int it = 0; it < text.size(); ++it)
+		for (int it = 0; it < static_cast<int>(text.size()); ++it)
 		{
 			char c = text[it];
 			if (c == '{')
